# openapi.yaml
openapi: 3.1.0
info:
  title: ZVERY Business Navigator API
  version: 0.1.0
  description: >
    Один контракт /api/v1. services/api слушает :8000 и отдаёт каталог.
    services/bot слушает :8001 и отдаёт пользовательские маршруты.
    Прямой :8000 не реализует bot-маршруты. GET /api/v1/measures/saved на API
    совпадает с карточкой ID saved и отвечает 404 measure_not_found.
servers:
  - url: http://127.0.0.1:8000
    description: services/api, прямой порт каталога
  - url: http://127.0.0.1:8001
    description: services/bot, прямой порт пользователя и webhook
  - url: https://zverybot.ru
    description: Публичный gateway. Маршрутизация этой фазой не меняется.
paths:
...
  /api/v1/catalog/filters:
    get:
      x-zvery-service: api
      operationId: catalogFilters
      responses:
        '200':
          description: Filter vocabulary and content gaps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogFilters'
        '500':
          description: Catalog failed schema validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/measures:
    get:
      x-zvery-service: api
      operationId: listMeasures
      responses:
        '200':
          description: Catalog records in file order. JSON array, not an items wrapper.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasureList'
        '500':
          description: Catalog failed schema validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/measures/{measure_id}:
    get:
      x-zvery-service: api
      operationId: getMeasure
      parameters:
        - $ref: '#/components/parameters/MeasureId'
      responses:
        '200':
          description: One catalog record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasureRecord'
        '404':
          description: Unknown id, including the literal id saved on the API port
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Catalog failed schema validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/recommendations:
    post:
      x-zvery-service: api
      operationId: recommendMeasures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: Matching catalog ids. Valid filters with no rows return items [].
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '422':
          description: Body or goal/sector outside the catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Catalog failed schema validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    MeasureId:
      in: path
      name: measure_id
      required: true
      schema:
        type: string
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              description: Создаёт сервер. Заголовок X-Request-ID клиента игнорируется.
          additionalProperties: false
      additionalProperties: false
...
    CatalogFilters:
      type: object
      required: [regions, roles, tax_modes, sectors, goals, accepted_tax_modes, content_gaps, catalog_version]
      properties:
        regions:
          type: array
          items:
            $ref: '#/components/schemas/Region'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        tax_modes:
          type: array
          description: Режимы, которые есть в записях, не весь enum.
          items:
            $ref: '#/components/schemas/TaxMode'
        sectors:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string
        accepted_tax_modes:
          type: array
          items:
            $ref: '#/components/schemas/TaxMode'
        content_gaps:
          type: array
          items:
            $ref: '#/components/schemas/Region'
        catalog_version:
          type: string
          const: demo-2026-09-19
    MeasureList:
      type: array
      items:
        $ref: '#/components/schemas/MeasureRecord'
    MeasureRecord:
      type: object
      required:
        - id
        - title
        - operator
        - region
        - roles
        - tax_modes
        - sector
        - goal
        - eligibility
        - documents
        - deadline
        - source_name
        - source_url
        - last_checked
        - freshness_status
        - data_status
        - disclaimer
      properties:
        id:
          type: string
        title:
          type: string
        operator:
          type: string
        region:
          $ref: '#/components/schemas/Region'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        tax_modes:
          type: array
          items:
            $ref: '#/components/schemas/TaxMode'
        sector:
          type: string
        goal:
          type: string
        eligibility:
          type: string
        documents:
          type: array
          items:
            type: string
        deadline:
          type: ['string', 'null']
        source_name:
          type: string
        source_url:
          type: string
        last_checked:
          type: string
        freshness_status:
          type: string
          enum: [fresh, reviewed, model]
        data_status:
          type: string
          enum: ['MODEL DATA', CONFIRMED]
        disclaimer:
          type: string
    RecommendationRequest:
      type: object
      required: [region, role, tax_mode]
      properties:
        region:
          $ref: '#/components/schemas/Region'
        role:
          $ref: '#/components/schemas/Role'
        tax_mode:
          $ref: '#/components/schemas/TaxMode'
        sector:
          type: string
        goal:
          type: string
    RecommendationResponse:
      type: object
      required: [items, catalog_version]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, title, data_status]
            properties:
              id:
                type: string
              title:
                type: string
              data_status:
                type: string
                enum: ['MODEL DATA', CONFIRMED]
        catalog_version:
          type: string
          const: demo-2026-09-19
...
    Region:
      type: string
      enum: [kazan, moscow, spb]
    Role:
      type: string
      enum: [ip, self_employed, llc]
    TaxMode:
      type: string
      enum: [npd, usn6, usn15, ausn, osno, none]
      description: npd не равен none. ausn не равен usn6 или usn15.